fix: seed emulator inbox during workspace setup
